feat: add integration test
The integration tests are designed to test the `azure-pipelines-orchestrator` on an actual Kubernetes cluster against a live Azure DevOps repo. During the test, a [kind cluster](https://kind.sigs.k8s.io/) is provisioned using the `kind-config.yaml` file that has both a local docker registry, and an `nginx` Ingress. The orchestrator is deployed, and a temporary pipeline is created for the life of tht test. Then, a pipeline job is queued to run and the script will monitor to see if the orchestrator correctly spawns a corresponding `Job` object in K8s.